requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
albumentations>=1.3.0
ultralytics>=8.0.0  # For YOLOv8
effdet>=0.4.0  # For EfficientDet
tensorboard>=2.13.0
tqdm>=4.65.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
timm>=0.9.0  # For EfficientDet
torchmetrics>=1.0.0